reorganized folder utils
