+++ subgraph/tsconfig.json
@@ -1,4 +1,9 @@
 {
   "extends": "./node_modules/assemblyscript/std/assembly.json",
-  "include": ["./src/**/*.ts"]
+  "include": ["./src/**/*.ts"],
+  "compilerOptions": {
+    "paths": {
+      "generated": ["./generated"]
+    }
+  }
 }
